ghetto initial tnetstring/otnetstring support.
added TSICTString lib, which might not be the fastest implementation (due to
how often data gets shuffled around), but was reasonably simple and is fast
enough not to care for this use case. perhaps updating to use reverse-string
concatenation voodoo would help, but not enough for me to care.

made xcode/clang stop bitching about two unused function parameters.