Update RSpec to v3.6
I made the minimal number of changes to the specs to make the work in
v3. The specs still need some additional updates to bring them up to
date with the latest RSpec style. The changes are:
* update truth checking matchers
* remove color_enable=true from the rake task because it is now the
  default behavior

The latest version of rspec also resolves a path issue that I was
encountering when trying to run the specs from within Rake.